+++ cvat-canvas/src/typescript/interactionHandler.ts
@@ -17,38 +17,25 @@ export interface InteractionHandler {
     transform(geometry: Geometry): void;
     interact(interactData: InteractionData): void;
     configurate(config: Configuration): void;
+    destroy(): void;
     cancel(): void;
 }
 
 export class InteractionHandlerImpl implements InteractionHandler {
     private onInteraction: (shapes: InteractionResult[] | null, shapesUpdated?: boolean, isDone?: boolean) => void;
-
     private configuration: Configuration;
-
     private geometry: Geometry;
-
     private canvas: SVG.Container;
-
     private interactionData: InteractionData;
-
     private cursorPosition: { x: number; y: number };
-
     private shapesWereUpdated: boolean;
-
     private interactionShapes: SVG.Shape[];
-
     private currentInteractionShape: SVG.Shape | null;
-
     private crosshair: Crosshair;
-
     private threshold: SVG.Rect | null;
-
     private thresholdRectSize: number;
-
     private intermediateShape: PropType<InteractionData, 'intermediateShape'>;
-
     private drawnIntermediateShape: SVG.Shape;
-
     private thresholdWasModified: boolean;
 
     private prepareResult(): InteractionResult[] {
@@ -473,13 +460,8 @@ export class InteractionHandlerImpl implements InteractionHandler {
             }
         });
 
-        window.addEventListener('keyup', this.onKeyUp);
-        window.addEventListener('keydown', this.onKeyDown);
-
-        this.canvas.on('destroy.canvas', ():void => {
-            window.removeEventListener('keyup', this.onKeyUp);
-            window.removeEventListener('keydown', this.onKeyDown);
-        });
+        window.document.addEventListener('keyup', this.onKeyUp);
+        window.document.addEventListener('keydown', this.onKeyDown);
     }
 
     public transform(geometry: Geometry): void {
@@ -559,5 +541,10 @@ export class InteractionHandlerImpl implements InteractionHandler {
     public cancel(): void {
         this.release();
         this.onInteraction(null);
+    }
+
+    public destroy(): void {
+        window.document.removeEventListener('keyup', this.onKeyUp);
+        window.document.removeEventListener('keydown', this.onKeyDown);
     }
 }