+++ cvat-canvas/src/typescript/svg.patch.ts
@@ -167,18 +167,23 @@ SVG.Element.prototype.resize = function constructor(...args: any): any {
         handler = this.remember('_resizeHandler');
         handler.resize = function (e: any) {
             const { event } = e.detail;
+            this.rotationPointPressed = e.type === 'rot';
             if (
                 event.button === 0 &&
-                // ignore shift key for cuboid change perspective
-                (!event.shiftKey || this.el.parent().hasClass('cvat_canvas_shape_cuboid')) &&
-                !event.altKey
+                // ignore shift key for cuboids (change perspective) and rectangles (precise rotation)
+                (!event.shiftKey || (
+                    this.el.parent().hasClass('cvat_canvas_shape_cuboid')
+                    || this.el.type  === 'rect')
+                ) && !event.altKey
             ) {
                 return handler.constructor.prototype.resize.call(this, e);
             }
         };
         handler.update = function (e: any) {
-            this.m = this.el.node.getScreenCTM().inverse();
-            return handler.constructor.prototype.update.call(this, e);
+            if (!this.rotationPointPressed) {
+                this.m = this.el.node.getScreenCTM().inverse();
+            }
+            handler.constructor.prototype.update.call(this, e);
         };
     } else {
         originalResize.call(this, ...args);